move args to cmd/mod.rs
It makes more sense to organize this module by subcommand and put the
args into the top-level mod.rs.
