+++ src/cmd/mod.rs
@@ -1,11 +1,33 @@
 mod args;
 
-pub(crate) use self::args::{Args, Subcommand};
+use crate::cmd::args::BuildCommand;
 use anyhow::Result;
+use clap::Parser;
 use env_logger::Builder;
 use log::LevelFilter;
 
 const DEFAULT_LEVEL_FILTER: LevelFilter = LevelFilter::Warn;
+
+/// A tool for building custom variants of Bottlerocket.
+#[derive(Debug, Parser)]
+#[clap(about, long_about = None)]
+pub(crate) struct Args {
+    /// Set the logging level. One of [off|error|warn|info|debug|trace]. Defaults to warn. You can
+    /// also leave this unset and use the RUST_LOG env variable. See
+    /// https://github.com/rust-cli/env_logger/
+    #[clap(long = "log-level")]
+    pub(crate) log_level: Option<LevelFilter>,
+
+    #[clap(subcommand)]
+    pub(crate) subcommand: Subcommand,
+}
+
+#[derive(Debug, Parser)]
+pub(crate) enum Subcommand {
+    /// Build something, such as a Bottlerocket image or a kit of packages.
+    #[clap(subcommand)]
+    Build(BuildCommand),
+}
 
 /// Entrypoint for the `twoliter` command line program.
 pub(super) async fn run(args: Args) -> Result<()> {
